Using the right timezone with datetime
Ensured that we're using the UTC timezone whenever requesting the current
date/time from the system. The HA OS normally has its time zone set to
UTC, but by specifying UTC in datetime.now() calls, we will get a timezone
offset from the isoformat() calls. If we see "+00:00" as offset, we
replace it with "Z" that ISO8601 allows for.