Make lots of improvements to Service
General clean up and removal of dead or unnecessary code.

Optimized the raising of the update event to only do it once for status
messages. As it was before we would get 2 updates, one for the request
and one for the response. Since we'll always get 2 updates for status
requests, we ignore the first one and only raise the event once for the
response.