Fix ping rate handling
Reworked ping rate handling to correctly reset the ping rates on a per
dataset type basis once we forward a request for that type.

Official ping rates coming from the service are now cached during
response handling for the Status dataset type.

Ping rates are now counted up (to the cached value obtained from the
service) instead of counted down to 0 to simplify the logic a bit. Also,
ping rates have their own last_updated field to allow for the correct
calculation of the elapsed time.

Also included in this change:

- Bumped version to 0.2.2
- Fixed pyproxy imports to rely only on the top level module itself (the
  pyproxy package was fixed to allow for this)